+++ src/CleanRegex/Internal/Prepared/Figure/InjectFigures.php
@@ -2,7 +2,6 @@
 namespace TRegx\CleanRegex\Internal\Prepared\Figure;
 
 use TRegx\CleanRegex\Internal\InvalidArgument;
-use TRegx\CleanRegex\Internal\Prepared\Template\AlternationToken;
 use TRegx\CleanRegex\Internal\Prepared\Template\LiteralToken;
 use TRegx\CleanRegex\Internal\Prepared\Template\Token;
 use TRegx\CleanRegex\Internal\Type\ValueType;
@@ -10,7 +9,7 @@ use UnderflowException;
 
 class InjectFigures implements CountedFigures
 {
-    /** @var array */
+    /** @var string[] */
     private $figures;
 
     public function __construct(array $figures)
@@ -20,26 +19,27 @@ class InjectFigures implements CountedFigures
 
     public function nextToken(): Token
     {
-        [$key, $value] = $this->nextEntry();
-
-        if (\is_string($value)) {
-            return new LiteralToken($value);
-        }
-        if (\is_array($value)) {
-            return new AlternationToken($value);
-        }
-        throw InvalidArgument::typeGiven("Invalid inject figure type. Expected string", new ValueType($value));
+        return new LiteralToken($this->nextString());
     }
 
-    private function nextEntry(): array
+    private function nextString(): string
+    {
+        $figure = $this->nextFigure();
+        if (\is_string($figure)) {
+            return $figure;
+        }
+        throw InvalidArgument::typeGiven("Invalid inject figure type. Expected string", new ValueType($figure));
+    }
+
+    private function nextFigure()
     {
         $key = \key($this->figures);
         if ($key === null) {
             throw new UnderflowException();
         }
-        $value = \current($this->figures);
+        $item = \current($this->figures);
         \next($this->figures);
-        return [$key, $value];
+        return $item;
     }
 
     public function count(): int